backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langgraph==0.0.20
chromadb==0.4.18
pydantic-settings==2.1.0
tiktoken==0.5.2
openai==1.6.1
azure-ai-inference==1.0.0b1
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
python-magic==0.4.27
requests==2.31.0
numpy==1.26.2
pandas==2.1.4
